\name{estDem-package}
\alias{estDem-package}
\alias{estDem}
\docType{package}
\title{
\packageTitle{estDem}
}
\description{
\packageDescription{estDem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{estDem}
\packageIndices{estDem}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{estDem}

Maintainer: \packageMaintainer{estDem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
