\name{EstDemand-package}
\alias{EstDemand-package}
\alias{EstDemand}
\docType{package}
\title{
\packageTitle{EstDemand}
}
\description{
\packageDescription{EstDemand}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EstDemand}
\packageIndices{EstDemand}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{EstDemand}

Maintainer: \packageMaintainer{EstDemand}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
